REFACTOR: Move AST transformation helpers to separate module.
